fix: 진척도 반환 시 정수로 반환 수정

+++ src/main/java/com/amcamp/domain/sprint/dao/SprintRepositoryImpl.java
@@ -34,7 +34,7 @@ public class SprintRepositoryImpl implements SprintRepositoryCustom {
                                         sprint.toDoInfo.startDt,
                                         sprint.toDoInfo.dueDt,
                                         sprint.toDoInfo.toDoStatus,
-                                        sprint.progress))
+                                        sprint.progress.intValue()))
                         .from(sprint)
                         .where(lastSprintId(lastSprintId), sprint.project.id.eq(projectId))
                         .orderBy(sprint.title.asc())
