+++ src/main/java/com/amcamp/domain/sprint/dto/response/SprintInfoResponse.java
@@ -12,7 +12,7 @@ public record SprintInfoResponse(
         @Schema(description = "스프린트 시작 날짜", defaultValue = "2026-02-01") LocalDate startDt,
         @Schema(description = "스프린트 마감 날짜", defaultValue = "2026-03-01") LocalDate dueDt,
         @Schema(description = "스프린트 진행 상태", defaultValue = "NOT_STARTED") ToDoStatus status,
-        @Schema(description = "스프린트 진척도", defaultValue = "0") Double progress) {
+        @Schema(description = "스프린트 진척도", defaultValue = "0") Integer progress) {
     public static SprintInfoResponse from(Sprint sprint) {
         return new SprintInfoResponse(
                 sprint.getId(),
@@ -21,6 +21,6 @@ public record SprintInfoResponse(
                 sprint.getToDoInfo().getStartDt(),
                 sprint.getToDoInfo().getDueDt(),
                 sprint.getToDoInfo().getToDoStatus(),
-                sprint.getProgress());
+                sprint.getProgress().intValue());
     }
 }
